Rearchitect function placement to hint at OO code structure
