+++ scripts/rallycron.php
@@ -1,9 +1,5 @@
 <?php
-require('include/curl.php');
-require('include/slack.php');
-require('include/rally.php');
-require('include/rallyme.config.php');
-require('include/slack.config.php');
+require('include/rallycron.inc.php');
 
 date_default_timezone_set('UTC');
